+++ spec/write_json_action_spec.rb
@@ -1,10 +1,10 @@
 require "fileutils"
 
 describe "write_json should" do
-  FILE_DIR = "#{__dir__}/resources/tmp"
+  let(:file_dir) { "#{__dir__}/resources/tmp" }
 
   after :each do
-    FileUtils.rm_rf(FILE_DIR) if File.directory?(FILE_DIR)
+    FileUtils.rm_rf(file_dir) if File.directory?(file_dir)
   end
 
   it "save a json file containing the hash values in the provided path" do
@@ -17,29 +17,33 @@ describe "write_json should" do
       ]
     }
 
-    expect(File.exist?("#{FILE_DIR}/my_json.json"))
+    expect(File.exist?("#{file_dir}/my_json.json"))
 
     Fastlane::Actions::WriteJsonAction.run({
-      file_path: "#{FILE_DIR}/my_json.json",
+      file_path: "#{file_dir}/my_json.json",
       hash: hash_value
     })
   end
 
   it "raise an error if hash: value is nil" do
-    expect(Fastlane::UI).to receive(:user_error!)
+    expect(FastlaneCore::UI).to receive(:user_error!)
 
-    Fastlane::Actions::WriteJsonAction.run({
-      file_path: "#{FILE_DIR}/my_json.json",
-      hash: nil
-    })
+    expect do
+      Fastlane::Actions::WriteJsonAction.run({
+        file_path: "#{file_dir}/my_json.json",
+        hash: nil
+      })
+    end.to raise_error(StandardError)
   end
 
   it "raise an error if file_path: value is nil" do
-    expect(Fastlane::UI).to receive(:user_error!)
+    expect(FastlaneCore::UI).to receive(:user_error!)
 
-    Fastlane::Actions::WriteJsonAction.run({
-      file_path: nil,
-      hash: {}
-    })
+    expect do
+      Fastlane::Actions::WriteJsonAction.run({
+        file_path: nil,
+        hash: {}
+      })
+    end.to raise_error(StandardError)
   end
 end